Move dto to separate crate
