+++ Cargo.toml
@@ -5,7 +5,7 @@ edition = "2021"
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 [workspace]
-members = [".", "db/postgres/entity", "db/postgres/migration", "db/mongo/collection", "db/repo", "db/mongo/schema", "message_broker"]
+members = [".", "db/postgres/entity", "db/postgres/migration", "db/mongo/collection", "db/repo", "db/mongo/schema", "message_broker", "dto"]
 
 [dependencies]
 entity = { path = "db/postgres/entity" }
@@ -14,6 +14,7 @@ repo = {path = "db/repo"}
 schema = {path = "db/mongo/schema"}
 collection = {path = "db/mongo/collection"}
 message_broker = {path = "message_broker"}
+dto = {path = "dto"}
 
 axum = { version = "0.7.5", features = ["macros"] }
 tokio = { version = "1.36.0", features = ["full"] }
