+++ db/repo/src/lib.rs
@@ -1,3 +1,2 @@
 pub mod dao;
-pub mod dto;
 pub mod utils;
