+++ db/repo/src/utils/mod.rs
@@ -1,4 +1,3 @@
-pub mod dto;
 pub mod repository;
 pub mod user;
 pub mod user_repo;
