+++ db/repo/src/utils/user/user_test_helper.rs
@@ -5,7 +5,7 @@ use collection::user::{TestUserCollection, User};
 use collection::MongoCollection;
 
 use crate::dao::user_repo::UserRepository;
-use crate::dto::user_dto::{CreateUserDto, UpdateUserDto, UserDto};
+use dto::user_dto::{CreateUserDto, UpdateUserDto, UserDto};
 
 pub fn get_mock_repo() -> UserRepository {
     let collection: Arc<dyn MongoCollection<User>> = Arc::new(TestUserCollection::default());
