+++ tests/api/user_repo_info_api.rs
@@ -5,9 +5,9 @@ use serial_test::serial;
 use uuid::Uuid;
 
 use collection::user_repo_info::UserRepoInfoOperation;
-use ia_11_vorobei_ant::web::dto::user_repo_dto::{OneToManyDto, OneToOneDto};
-use repo::dto::DtoList;
-use repo::dto::user_repo_info_dto::UserRepoInfoDto;
+use dto::{OneToManyDto, OneToOneDto};
+use dto::DtoList;
+use dto::user_repo_info_dto::UserRepoInfoDto;
 
 use crate::common::Setup;
 use crate::helpers::user_repo_api_helper;
